+++ Sphinx/ChangeLog.rst
@@ -1,6 +1,8 @@
 ChangeLog
 =========
-* April 18, 2015
+* April 25, 2015
+
+  - Added ``example/stop_example.py``.
 
   - Corrected ``test/test_sound.py``.